Add a check for a blank message body
SNS will throw a bad argument error if the body of a notification is empty. Just a blank line is enough to keep it happy. This change changes a blank body to a single blank line.

This is useful for SMS notifications, where SNS only delivers the subject as the SMS, so the body of the message has no purpose and can be blank.